refactor(delibs): refactor to use custom react hooks for readability
